+++ client/assets/scripts/vue.js
@@ -30,6 +30,38 @@ new Vue({
                 desc: 'Pesawatnya kevin rusak lagi',
                 author: 'cecha'
             }
-        ]
+        ],
+        image: [],
+        formData: new FormData(),
+    },
+    methods:{
+        uploadtest:function(){
+            if(this.image.length > 0){
+                for(let i = 0; i < this.image.length; i++){
+                    let file = this.image[i];
+                    this.formData.append('image',file, file.name)
+                }
+            //     var formData = new FormData();
+
+            // for (var i = 0; i < files.length; i++) {
+            //     var file = files[i];
+            //     formData.append('pictures', file, file.name);
+            // }
+            }
+            console.log(this.formData)
+            
+            // this.formData.set('image', this.image)
+            // console.log(this.formData)
+            axios.post('http://localhost:3000/upload',this.formData)
+        },
+        ambilFileNya: function(event){
+            let img = event.target.files
+            console.log(img)
+            Object.keys(img).forEach(value => {
+                this.image.push(img[value])
+            })
+           
+            console.log(this.image)
+        }
     }
 })
