+++ server/routes/login.routes.js
@@ -1,18 +1,12 @@
 const express = require('express')
 const router = express.Router()
-const { logIn, readAllUsers, destroyUser } = require('../controllers/login.routes')
+// const { logIn, readAllUsers, destroyUser } = require('../controllers/login.routes')
 
-// router.get('/', (req, res) => {
-//   res.status(201).json({
-//     message: 'login!'
-//   })
-// })
+// router.post('/', logIn)
+// router.get('/', readAllUsers)
+// // router.delete('/:id', destroyUser)
 
-router.post('/', logIn)
-router.get('/', readAllUsers)
-// router.delete('/:id', destroyUser)
-
-const { createUser, loginUser, deleteUser, getUser } = require('../controllers/login.routes')
+const { createUser, loginUser, deleteUser, getUser } = require('../controllers/login.controller')
 
 router.post('/create',createUser)
 router.post('/login', loginUser)
